# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AICSS Backend",
    "description": "AI Cinematic Spatial System — Depth + Segmentation + Spatial Layers",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/aicss/v2/meshes/{mesh_id}/download": {
      "get": {
        "tags": [
          "v2 - 3D Mesh Export",
          "v2 - 3D Mesh Export"
        ],
        "summary": "Api Download Mesh",
        "description": "下载指定的 mesh 文件。\n\n支持 GLB 和 FBX 格式。",
        "operationId": "api_download_mesh_api_aicss_v2_meshes__mesh_id__download_get",
        "parameters": [
          {
            "name": "mesh_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mesh Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}